+++ artemis/tests/auvergne_louis_01_test.py
@@ -1,6 +1,6 @@
-import base_pytest2
+import base_pytest
 
-class TestAuvergne(base_pytest2.TestFixture):
+class TestAuvergne(base_pytest.TestFixture):
 
     """
     test for new_default with data from auvergne
@@ -10,7 +10,7 @@ class TestAuvergne(base_pytest2.TestFixture):
         http://jira.canaltp.fr/browse/NAVITIAII-2016
         """
 
-        base_pytest2.journey_Test(self, _from="admin:fr:63135",
+        base_pytest.journey_Test(self, _from="admin:fr:63135",
                     to="3.121833801269531;45.885276435738504", datetime="20160118T120300",
                     first_section_mode=['bike', 'bss', 'walking', 'car'],
                     last_section_mode=['walking'],
